backend/requirements.txt

celery>=5.4.0
django>=5.1.7
django-cors-headers>=4.7.0
django-extensions>=3.2.3
django-filter>=25.1
django-leaflet>=0.31.0
djangorestframework>=3.15.2
djangorestframework-gis>=1.0
djangorestframework-simplejwt>=5.5.0
dotenv>=0.9.9
drf-yasg>=1.21.10
gunicorn>=23.0.0
pillow>=11.1.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
redis>=5.2.1
